Votes-tests
Updated Blog tests to include Votes, and they all pass , except for 3 comment tests, 1 RemoveVote test (the test excpects a JSON response after deleting a vote but the controller is returning 204 noContent, which is correct), so it need a small fix but im currently sick and can't think properly so will fix in a forseeable future, and 2 GetBlog tests which are getting a 405 (method not allowed) meaning either the method is wrong or the auth token is expired/invalid because the endpoind does exist in the controller and should be correct, so it's either blog id doesnt exist, messed inside of the data since I tested the features in Swagger and was messing with the data, or what I said above.
Also, updated BlogProfile.cs to now change Status into a string from an enum since there was a type mismatch error in the tests.